+++ service/customerprofiles/doc.go
@@ -3,18 +3,13 @@
 // Package customerprofiles provides the client and types for making API
 // requests to Amazon Connect Customer Profiles.
 //
-// Welcome to the Amazon Connect Customer Profiles API Reference. This guide
-// provides information about the Amazon Connect Customer Profiles API, including
-// supported operations, data types, parameters, and schemas.
-//
 // Amazon Connect Customer Profiles is a unified customer profile for your contact
 // center that has pre-built connectors powered by AppFlow that make it easy
 // to combine customer information from third party applications, such as Salesforce
 // (CRM), ServiceNow (ITSM), and your enterprise resource planning (ERP), with
-// contact history from your Amazon Connect contact center.
-//
-// If you're new to Amazon Connect , you might find it helpful to also review
-// the Amazon Connect Administrator Guide (https://docs.aws.amazon.com/connect/latest/adminguide/what-is-amazon-connect.html).
+// contact history from your Amazon Connect contact center. If you're new to
+// Amazon Connect, you might find it helpful to review the Amazon Connect Administrator
+// Guide (https://docs.aws.amazon.com/connect/latest/adminguide/).
 //
 // See https://docs.aws.amazon.com/goto/WebAPI/customer-profiles-2020-08-15 for more information on this service.
 //