+++ service/xray/api.go
@@ -3658,20 +3658,31 @@ func (s *DeleteSamplingRuleOutput) SetSamplingRuleRecord(v *SamplingRuleRecord) 
 	return s
 }
 
-// Information about a connection between two services.
+// Information about a connection between two services. An edge can be a synchronous
+// connection, such as typical call between client and service, or an asynchronous
+// link, such as a Lambda function which retrieves an event from an SNS queue.
 type Edge struct {
 	_ struct{} `type:"structure"`
 
 	// Aliases for the edge.
 	Aliases []*Alias `type:"list"`
 
+	// Describes an asynchronous connection, with a value of link.
+	EdgeType *string `type:"string"`
+
 	// The end time of the last segment on the edge.
 	EndTime *time.Time `type:"timestamp"`
+
+	// A histogram that maps the spread of event age when received by consumers.
+	// Age is calculated each time an event is received. Only populated when EdgeType
+	// is link.
+	ReceivedEventAgeHistogram []*HistogramEntry `type:"list"`
 
 	// Identifier of the edge. Unique within a service map.
 	ReferenceId *int64 `type:"integer"`
 
-	// A histogram that maps the spread of client response times on an edge.
+	// A histogram that maps the spread of client response times on an edge. Only
+	// populated for synchronous edges.
 	ResponseTimeHistogram []*HistogramEntry `type:"list"`
 
 	// The start time of the first segment on the edge.
@@ -3705,9 +3716,21 @@ func (s *Edge) SetAliases(v []*Alias) *Edge {
 	return s
 }
 
+// SetEdgeType sets the EdgeType field's value.
+func (s *Edge) SetEdgeType(v string) *Edge {
+	s.EdgeType = &v
+	return s
+}
+
 // SetEndTime sets the EndTime field's value.
 func (s *Edge) SetEndTime(v time.Time) *Edge {
 	s.EndTime = &v
+	return s
+}
+
+// SetReceivedEventAgeHistogram sets the ReceivedEventAgeHistogram field's value.
+func (s *Edge) SetReceivedEventAgeHistogram(v []*HistogramEntry) *Edge {
+	s.ReceivedEventAgeHistogram = v
 	return s
 }
 
@@ -9143,9 +9166,9 @@ type Trace struct {
 	// and subsegments.
 	Id *string `min:"1" type:"string"`
 
-	// LimitExceeded is set to true when the trace has exceeded one of the defined
-	// quotas. For more information about quotas, see Amazon Web Services X-Ray
-	// endpoints and quotas (https://docs.aws.amazon.com/general/latest/gr/xray.html).
+	// LimitExceeded is set to true when the trace has exceeded the Trace document
+	// size limit. For more information about this limit and other X-Ray limits
+	// and quotas, see Amazon Web Services X-Ray endpoints and quotas (https://docs.aws.amazon.com/general/latest/gr/xray.html).
 	LimitExceeded *bool `type:"boolean"`
 
 	// Segment documents for the segments and subsegments that comprise the trace.